+++ app/client/doc/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.33] - 2023-06-09
+## [0.0.0.34] - 2023-06-12
 
 - Exported `rootCmd` to support cross-package usage (i.e. subcommands w/ own pkgs)
 - Refactored common CLI code
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Replaced RPC_HOST with POCKET_REMOTE_CLI_URL where appropriate
 - Added support for overriding of `remote-cli-url` flag with env var
 - Improve error handling in client CLI
+
+## [0.0.0.33] - 2023-06-12
+
+- Renamed `NewRPCPeerstoreProvider()` and `NewPersistencePeerstoreProvider()` to `Create()` (per package)
 
 ## [0.0.0.32] - 2023-05-25