+++ p2p/CHANGELOG.md
@@ -7,16 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.55] - 2023-06-09
+## [0.0.0.55] - 2023-06-12
 
 - Replaced `RPC_HOST` with `POCKET_REMOTE_CLI_URL` or `--pocket-remote-cli-url` where appropriate
 
-## [0.0.0.54] - 2023-06-08
+## [0.0.0.54] - 2023-06-12
 
 - Replaced embedded `modules.Module` with simpler `modules.IntegratableModule` in `PeerstoreProvider` interface
 - Removed unused `PeerstoreProvider#GetP2PConfig()` method
 - Added `PeerstoreProvider#GetUnstakedPeerstore()` method
 - Added temporary `unstakedPeerstoreProvider` interface
+- Renamed `NewRPCPeerstoreProvider()` and `NewPersistencePeerstoreProvider()` to `Create()` (per package)
+- Updated `persistencePeerstoreProvider` to retrieve all staked actors (not just validators)
 
 ## [0.0.0.53] - 2023-06-01