+++ shared/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.60] - 2023-06-08
+## [0.0.0.60] - 2023-06-12
 
 - Added `Factory` generic type
 - Improved godoc comments